+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - **Text based controls**
   - _Sometimes weird blink & poor render during the animation on focus_
   - _Button clear changes control size on hover if parent is flexbox_
+- [SelectControl](src/controls/select.ts). [SelectManyControl](src/controls/selectMany.ts). _storageKey + items as Promise don't work together_
 
 ### BREAKING CHANGES